Fixes the issue on polymorphic types after pattern matching compilation
